Consolidated common modules and removed unused activity in manifest to address lint error
