+++ app/src/main/java/com/atomicrobot/carbon/app/Modules.kt
@@ -1,8 +1,7 @@
-package com.atomicrobot.carbon.data
+package com.atomicrobot.carbon.app
 
 import android.content.Context
 import androidx.annotation.VisibleForTesting
-import com.atomicrobot.carbon.app.Settings
 import com.atomicrobot.carbon.data.api.github.GitHubApiService
 import com.atomicrobot.carbon.data.api.github.GitHubInteractor
 import com.atomicrobot.carbon.deeplink.DeepLinkInteractor
@@ -23,12 +22,25 @@ import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
 import retrofit2.converter.moshi.MoshiConverterFactory
 import java.io.File
 
-interface OkHttpSecurityModifier {
-    fun apply(builder: OkHttpClient.Builder)
-}
+class Modules {
 
-class DataModule {
-    val dataModule = module {
+    companion object {
+        private const val LOADING_DELAY_MS: Long = 500
+        private const val DISK_CACHE_SIZE = 50 * 1024 * 1024 // 50MB
+        const val BASE_URL = "baseUrl"
+    }
+
+    val appModules = module {
+        single {
+            Settings(context = androidContext())
+        }
+
+        single(named("loading_delay_ms")) {
+            LOADING_DELAY_MS
+        }
+    }
+
+    val dataModules = module {
         single {
             val cacheDir = File(androidApplication().cacheDir, "http")
             Cache(cacheDir, DISK_CACHE_SIZE.toLong())
@@ -76,7 +88,9 @@ class DataModule {
         single {
             DeepLinkInteractor()
         }
+    }
 
+    val viewModelModules = module {
         viewModel {
             SplashViewModel(
                 app = androidApplication(),
@@ -129,10 +143,8 @@ class DataModule {
     ): GitHubInteractor {
         return GitHubInteractor(context, api)
     }
+}
 
-    companion object {
-        private const val DISK_CACHE_SIZE = 50 * 1024 * 1024 // 50MB
-
-        const val BASE_URL = "baseUrl"
-    }
+interface OkHttpSecurityModifier {
+    fun apply(builder: OkHttpClient.Builder)
 }
